refactor: Move svgs to /icons
