+++ src/components/Form/Checkbox/Checkbox.tsx
@@ -1,8 +1,8 @@
 
 import React from "react"
 import { CheckboxStyled } from "./CheckboxStyles"
-import { ReactComponent as TickIcon } from "../../../assets/tick.svg"
-import { ReactComponent as MinusIcon } from "../../../assets/minus.svg"
+import { ReactComponent as TickIcon } from "../../../assets/icons/tick.svg"
+import { ReactComponent as MinusIcon } from "../../../assets/icons/minus.svg"
 import { renderErrorMessage } from "../FormCommon"
 import { TextAlignKeys, TextSizeKeys, TextWeightKeys } from "../../../styles/themeHelpers"
