+++ src/components/Form/Input/InputPassword.tsx
@@ -2,8 +2,8 @@
 import React from "react"
 import { InputStyled } from "./InputStyles"
 import { useState } from "react"
-import { ReactComponent as HideIcon } from "../../../assets/not_preview.svg"
-import { ReactComponent as ShowIcon } from "../../../assets/preview.svg"
+import { ReactComponent as HideIcon } from "../../../assets/icons/not_preview.svg"
+import { ReactComponent as ShowIcon } from "../../../assets/icons/preview.svg"
 import { renderErrorMessage } from "../FormCommon"
 import { InputCommonProps } from "./InputCommon"
 import Spinner, { SpinnerSize } from "../../Spinner/Spinner"
